Use dep.exists() to check for presence.
`semver`'s `satisfies` has somewhat bizarre issues around matching
prerelease versions. In this case, we don't _really_ care what version
you are using at all, we just want to check if you have a given
dependency installed.

This updates to using a new API in ember-cli-version-checker to simply
check if a given dep exists or not.